Start to decompose Main page
Separate Containers vue to main page
