+++ src/components/Containers.vue
@@ -1,0 +1,13 @@
+<template>
+<div class="main">
+  <div class="menubar">
+    <a href="#" class="btn btn-sq-sm btn-primary"><i class="fa fa-plus fa-2x"></i><br/>Ajouter</a><span class="vr"></span><a href="#" class="btn btn-sq-sm btn-primary"><i class="fa fa-play fa-2x"></i><br/>Démarrer</a><a href="#" class="btn btn-sq-sm btn-primary-stop"><i class="fa fa-stop fa-2x"></i><br/>Arrêter</a><a href="#" class="btn btn-sq-sm btn-primary"><i class="fa fa-refresh fa-2x"></i><br/>Relancer</a><a href="#" class="btn btn-sq-sm btn-primary"><i class="fa fa-minus fa-2x"></i><br/>Supprimer</a>
+  </div>
+  <div class="section_title">
+    <span class="section_title_text">Gestion des applications</span><br/>
+    <span class="section_title_subtext">Ici, installez, supprimez ou ajoutez de nouvelles application à votre saedbox</span>
+  </div>
+
+
+</div>
+</template>
